requirements.txt
absl-py==1.3.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
anyio==3.6.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bentoml==1.0.5
blinker==1.4
Brlapi==0.7.0
build==0.8.0
cachetools==5.2.0
cattrs==22.1.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
circus==0.17.1
Click==7.0
cloudpickle==2.2.0
cmake==3.24.1.1
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
contextlib2==21.6.0
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
datasets==2.4.0
dbus-python==1.2.16
debugpy==1.6.0
decorator==4.4.2
deepmerge==1.0.1
defer==1.0.6
Deprecated==1.2.13
dill==0.3.5.1
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.0.0rc9
executing==0.8.3
fastapi==0.85.0
fasteners==0.14.1
filelock==3.8.0
flatbuffers==22.10.26
fonttools==4.33.3
frozenlist==1.3.1
fs==2.4.16
fsspec==2022.8.2
future==0.18.2
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.50.0
h11==0.13.0
h5py==3.7.0
html5lib==1.0.1
httplib2==0.14.0
huggingface-hub==0.9.1
idna==2.8
imageio==2.19.0
importlib-metadata==1.5.0
iniconfig==1.1.1
instaloader==4.9
ipykernel==6.13.0
ipython==8.2.0
ipython_genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.3.0
jupyter-core==4.10.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
monotonic==1.5
more-itertools==4.2.0
multidict==6.0.2
multiprocess==0.70.13
nest-asyncio==1.5.5
netifaces==0.10.4
networkx==2.8
numpy==1.22.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opentelemetry-api==1.12.0
opentelemetry-instrumentation==0.33b0
opentelemetry-instrumentation-aiohttp-client==0.33b0
opentelemetry-instrumentation-asgi==0.33b0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
opentelemetry-util-http==0.33b0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
paramiko==2.6.0
parso==0.8.3
pathspec==0.10.1
pep517==0.13.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.1.0
pip-tools==6.8.0
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==2.0.10
protobuf==3.19.6
psutil==5.9.0
pure-eval==0.2.2
py==1.11.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycups==1.9.73
pydantic==1.10.2
pyforest==1.1.0
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvml==11.4.1
pyparsing==3.0.8
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.5
pytest==7.1.3
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.1
PyWavelets==1.3.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qudida==0.0.4
reportlab==3.5.34
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
responses==0.18.0
rich==12.5.1
rsa==4.9
schema==0.7.5
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.5.0
simple-di==0.1.5
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==1.9.5
stack-data==0.2.0
starlette==0.20.4
systemd-python==234
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
threadpoolctl==3.1.0
tifffile==2022.4.28
timm==0.6.11
tomli==2.0.1
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
torchviz==0.0.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.3.0
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
uvicorn==0.18.3
wadllib==1.3.3
watchfiles==0.17.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
xkit==0.0.0
xxhash==3.0.0
yarl==1.8.1
zipp==1.0.0
